Use own icons for up/down buttons
